Starting to modify various int variables and functions to size_t

+++ src/brains/ForceField.hpp
@@ -132,7 +132,7 @@ namespace OpenMD {
       wildCardAtomTypeName_ = wildCard;
     }
     
-    unsigned int getNAtomType() {
+    size_t getNAtomType() {
       return atomTypeCont_.size();
     }
